+++ python/exp03_syr2k.py
@@ -12,9 +12,9 @@ def syr2k_implicit_noup(A, B, C):
 
 def exp03_syr2k(b, n):
 
-    A = torch.randn(n, n)
-    B = torch.randn(n, n)
-    C = torch.randn(n, n)
+    A = torch.randn((n, n), dtype=torch.float64)
+    B = torch.randn((n, n), dtype=torch.float64)
+    C = torch.randn((n, n), dtype=torch.float64)
     C = C + torch.transpose(C, 0, 1)
 
     res = b.benchmark('syr2k_implicit_noup', syr2k_implicit_noup, A, B, C)
